Copy dependencies required for painless before resource processing
This removes the jar files required for the painless plugin from the
repo and copies them before resources are processed.

This doesn't fix the ugliness in Server.java when coping the jars,
to make that at least a bit better and nearly only dependent on
configuration we strip the version when copying.

Further it converts indenting in pom.xml to spaces just as the rest of
the code.