make default language configurable
Adds a new command line parameter for choosing the default language
returned for queries. If not set the results will returned in the
locale form (aka the contents of the 'name' tag without language).

As a result of these changes it is now also possible to explicitly
ask for locale results with 'lang=default'.