evaluate extratags->linked_place from placex table
In recent versions of Nominatim the type of a linked place
has been moved from extratags->place to extratags->linked_place
so that it does not overwrite values set by mappers.

Change Photon to look at both fields. Keep a preference to the
'place' value as this would be whatever the mapper has chosen.

This change is backward compatible in the sense that Photon works
with older and newer versions of Nominatim.

Fixes #489.