simplify PhotonRequest
Unify FilteredPhotonRequest and PhotonRequest and get rid of the factory
methods for them. We can also get rid of a couple of generalisations
in the search request because of that. Move parsing of filter
expressions into PhotonRequest and add proper tests for the parser.

Remove the *RequestHandlerTests that only check for setting of internal
variables.

Remove unused ElasticsearchSearcherFactory.