basic framework for testing Nominatim connector
The framework uses an H2 in-memory database to provide the data.
Given that we only use simple SELECT statements, this is sufficiently
close to the actual usage. The only change required is for hstore
and geometry datatypes. hstore gets replaced by json. For geometry
use H2's own geometry types. Both types need a slightly different
form of marshalling. This is done by mocking the functions in
DBUtils. I haven't found a usable way to replace the static functions,
so DBUtils is now an instantiated class with member functions.