introduce enum for Photon's address type
The enum lists the possible address fields that Photon may index
and output. It is also responsible for mapping Nominatim's address
rank values to the address type.

PhotonDoc uses the type to define its address in a map instead of
having lots of separate fields. That simplifies code in a number
of places.

The commit also disallows adding an address part with the same level
as the place object itself.

Fixes #499.