only wait for yellow status on startup
The documentation says that the yellow status is reached, when the
primary shard is available and green status only when the replica
shards are online. Tests are run with a single shard only, thus
they can never reach green state because there is no replica shard
configured. Waiting for the green status in fact resulted in a
silent timeout of the operation after 30s. This made the API
tests annoyingly slow.

Having the primary shard available should be sufficient for most
operations. So this commit changes the status operation to only
wait for yellow state.