switch to a combined location bias/importance score
Switch to a simple linear decay function for weighing with importance.
It uses the importance as a direct weigh factor. The only modification
here is to ensure that the value range is above zero to avoid
cancelling out the score when multiplying the weight.

Location bias is switched to an exponential decay function with the
given distance in kilometers as the 0.8 weight limit. To ensure that
well-known places show up in the results nonetheless, use a second
weight function based on importance and use the max of the two
functions.