take geography into account for address computation
Use the more complex way of computing the address from Nominatim
for computing the rank 30 address. That makes sure that address
parts are not blindly taken from the parent but geographic
relations are taken into account. When streets go over boundaries,
all address fall then into the right side of the boundary.

Requires advanced SQL which is not supported by H2. Thus use the
previous simpler query for tests instead.