simplify handling of tag filter parameters
Tag filters are now handled as a list of simple filters instead of
being grouped by filter kind. That greatly simplifies the code when
parsing the incoming request and when building the query.

The ES query for excluded tags has changed to from
(not t1 and not t2 and not t3...) to not (t1 or t2 or t3).
This is equivalent and more simple to handle.