rework handling of document ID for multidocs
Instead of creating a document ID that consists of place ID and
housenumber, create an artificial sub ID for each document.
A single place ID is always updated as a batch. Thus if the sub IDs
are assigned sequentially, it is possible to find all documents belonging
to a given place ID by simply iterating over sub IDs until there is
no document anymore.