replace NominatimResult with document group generators
There are two specific generators now: one for splitting addresses,
one for splitting interpolations. As generators only omit document
iterators, we could in the future also generate sets that are not
related to address splitting. Furthermore it is now possible to
change other data in the PhotonDoc than housenumber. This is already
used in the split-address generator to correct the address of
place addresses.