properly shut down test instance
When running API tests, there is a server connection towards the
ES instance set up, which wasn't properly cleaned. That caused
oom errors at some point with the ES backend because the client
here seems to be rather heavy. It also means that the client is
expensive to shut down, so restructure tests to run as many tests as
possible against the same API frontend.